chore: make utils folder and move some files
